Add a new feature to allow users to toggle between minimal and full modes. This is done by hiding/showing the pane stack, menu actions, and toolbars. The main part of the process is handled in MainWindow::toggleViewMode(). The toggle view mode option appears in the View Menu and is associated with a new toolbar button. Two icons have been added in the resources. They come from: http://www.iconfinder.com/icondetails/42115/32/
In a working state, but the resizing of the main window and the audio file overview can be improved. Some menus or menu actions can be deleted/disabled in the minimal mode (although the changes are active in the hidden pane(s) and do not cause the application to crash). See TO-DO list in MainWindow::toggleViewMode().

In sonic-visualiser.pro: -lvamp-hostsdk has been added to the LIBS (the sonic-visualiser.pro file may be to be removed from the tracked files since the modification seems to depend on the system).